GET_Flask/requirements.txt
bcrypt==3.2.2
blinker==1.5
cachelib==0.10.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==4.4.2
dnspython==2.2.1
dwollav2==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
nulltype==2.3.1
Pillow==9.4.0
plaid-python==11.4.0
plivo==4.32.0
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.15
pytz==2022.7.1
ratelimit==2.2.1
redis==3.4.1
requests==2.27.1
rq==1.2.2
signnow-python-sdk==2.0.1
six==1.16.0
twilio==7.16.2
urllib3==1.26.14
Werkzeug==2.2.2
itsdangerous==2.1.2